Improve UIEntitySelection
